+++ packages/terra-button/tests/nightwatch/button-spec.js
@@ -1,5 +1,5 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
-
+// eslint-disable-next-line import/no-extraneous-dependencies
 const screenshot = require('terra-toolkit').screenshot;
 
 module.exports = {
@@ -14,117 +14,116 @@ module.exports = {
   'Displays a default button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/default`)
-      .assert.containsText('.terra-Button.terra-Button--default', 'Default');
+      .assert.containsText('#defaultButton', 'Default');
   },
 
   'Wraps long text at smaller screen sizes': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/long-text`)
-      .assert.containsText('.terra-Button.terra-Button--default', 'Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long');
+      .assert.containsText('#longTextButton', 'Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long');
   },
 
   'Displays a disabled button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/disabled`)
-      .assert.containsText('.terra-Button.terra-Button--default.is-disabled', 'Disabled Button');
+      .assert.containsText('#disabledButton', 'Disabled Button');
   },
 
   'Displays a disabled link with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/disabled-link`)
-      .assert.containsText('.terra-Button.terra-Button--link.is-disabled', 'Disabled Link');
+      .assert.containsText('#disabledLink', 'Disabled Link');
   },
 
   'Displays as a primary button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/primary`)
-      .assert.containsText('.terra-Button.terra-Button--primary', 'Primary');
+      .assert.containsText('#primaryButton', 'Primary');
   },
 
   'Displays as a secondary button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/secondary`)
-      .assert.containsText('.terra-Button.terra-Button--secondary', 'Secondary');
+      .assert.containsText('#secondaryButton', 'Secondary');
   },
 
   'Displays as a link with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/link`)
-      .assert.containsText('.terra-Button.terra-Button--link', 'Link');
+      .assert.containsText('#linkButton', 'Link');
   },
 
   'Displays as a tiny button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/tiny`)
-      .assert.containsText('.terra-Button.terra-Button--tiny', 'Tiny');
+      .assert.containsText('#tinyButton', 'Tiny');
   },
 
   'Displays as a small button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/small`)
-      .assert.containsText('.terra-Button.terra-Button--small', 'Small');
+      .assert.containsText('#smallButton', 'Small');
   },
 
   'Displays as a medium button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/medium`)
-      .assert.containsText('.terra-Button.terra-Button--medium', 'Medium');
+      .assert.containsText('#mediumButton', 'Medium');
   },
 
   'Displays as a large button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/large`)
-      .assert.containsText('.terra-Button.terra-Button--large', 'Large');
+      .assert.containsText('#largeButton', 'Large');
   },
 
   'Displays as a huge button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/huge`)
-      .assert.containsText('.terra-Button.terra-Button--huge', 'Huge');
+      .assert.containsText('#hugeButton', 'Huge');
   },
 
   'Displays as a button with the type set to button and with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/with-type-button`)
-      .assert.containsText('.terra-Button[type="button"]', 'Button Type is Default');
+      .assert.containsText('#buttonWithTypeButton', 'Button Type is Default');
   },
 
   'Displays as a button with the type set to reset and with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/with-type-reset`)
-      .assert.containsText('.terra-Button[type="reset"]', 'Button Type is Reset');
+      .assert.containsText('#buttonWithTypeReset', 'Button Type is Reset');
   },
 
   'Displays as a button with the type set to submit and with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/with-type-submit`)
-      .assert.containsText('.terra-Button[type="submit"]', 'Button Type is Submit');
+      .assert.containsText('#buttonWithTypeSubmit', 'Button Type is Submit');
   },
 
   'Displays an icon inline with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/icon-default`)
-      .assert.containsText('.terra-Button.terra-Button--default > :last-child', 'Button with Icon')
-      .assert.cssClassPresent('.terra-Button.terra-Button--default > :first-child', 'test');
+      .assert.containsText('#iconDefaultButton > :last-child', 'Button with Icon')
+      .assert.cssClassPresent('#iconDefaultButton > :first-child', 'test');
   },
 
   'Displays an icon only': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/icon-only`)
-      .assert.elementNotPresent('.terra-Button-text')
-      .assert.cssClassPresent('.terra-Button.terra-Button--default > :first-child', 'test');
+      .assert.cssClassPresent('#iconOnlyButton > :first-child', 'test');
   },
 
   'Displays a button with icons and text reversed': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/icon-reversed`)
-      .assert.containsText('.terra-Button.terra-Button--default > :first-child', 'Button with Icon and reversed')
-      .assert.cssClassPresent('.terra-Button.terra-Button--default > :last-child', 'test');
+      .assert.containsText('#iconReversedButton > :first-child', 'Button with Icon and reversed')
+      .assert.cssClassPresent('#iconReversedButton > :last-child', 'test');
   },
 
   'Displays as a compact button with the provided text': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/button-tests/compact`)
-      .assert.containsText('.terra-Button.terra-Button--compact', 'Compact');
+      .assert.containsText('#buttonCompact', 'Compact');
   },
 };