+++ packages/terra-status/tests/nightwatch/status-spec.js
@@ -1,5 +1,5 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
-
+// eslint-disable-next-line import/no-extraneous-dependencies
 const screenshot = require('terra-toolkit').screenshot;
 
 module.exports = {
@@ -14,28 +14,28 @@ module.exports = {
   'Displays a default status': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/status-tests/default`)
-      .waitForElementVisible('.terra-Status', 1000)
-      .expect.element('.terra-Status').text.to.contain('Sample text');
+      .waitForElementVisible('#status', 1000)
+      .expect.element('#status').text.to.contain('Sample text');
   },
   'Displays arrange status example': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/status-tests/arrange`)
-      .waitForElementVisible('.terra-Status', 1000)
-      .expect.element('.terra-Status').text.to.contain('Sample text');
+      .waitForElementVisible('#statusArrange', 1000)
+      .expect.element('#statusArrange').text.to.contain('Sample text');
   },
   'Displays icon status example': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/status-tests/icon`)
-      .waitForElementVisible('.terra-Status', 1000);
+      .waitForElementVisible('#statusIcon', 1000);
   },
   'Displays image status example': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/status-tests/image`)
-      .waitForElementVisible('.terra-Status', 1000);
+      .waitForElementVisible('#statusImage', 1000);
   },
   'Displays status with no style': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/status-tests/no-color`)
-      .waitForElementVisible('.terra-Status', 1000);
+      .waitForElementVisible('#statusNoColor', 1000);
   },
 };