+++ packages/terra-toggle-button/tests/nightwatch/toggle-button-spec.js
@@ -1,5 +1,5 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
-
+// eslint-disable-next-line import/no-extraneous-dependencies
 const screenshot = require('terra-toolkit').screenshot;
 
 module.exports = {
@@ -14,19 +14,19 @@ module.exports = {
   'Displays a default toggle-button': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/default`)
-      .assert.elementPresent('.terra-ToggleButton');
+      .assert.elementPresent('#toggleButton');
   },
 
   'Displays a toggle-button with customized button': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/customized-button`)
-      .assert.cssClassPresent('.terra-Button', 'terra-Button--link');
+      .assert.elementPresent('#linkButton');
   },
 
   'Displays a toggle-button with customized closedButtonText': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/closed-button-text`)
-      .assert.containsText('.terra-ToggleButton-buttonText', 'Custom Text');
+      .assert.containsText('#closedButtonText button > div > div:nth-child(2) > span', 'Custom Text');
   },
 
   'Displays a toggle-button with customized icon': (browser) => {
@@ -38,59 +38,59 @@ module.exports = {
   'Displays an animated toggle-button': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/animated`)
-      .assert.elementPresent('.terra-Toggle.is-animated');
+      .assert.elementPresent('#animatedToggle');
   },
 
   'Displays a toggle-button with animated icon': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/icon-animated`)
-      .assert.elementPresent('.terra-ToggleButton.is-icon-animated');
+      .assert.elementPresent('#animatedIcon');
   },
 
   'Displays an icon only toggle-button': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/icon-only`)
-      .assert.attributeContains('.terra-Button', 'aria-label', 'Custom Text')
-      .assert.containsText('.terra-Button', '');
+      .assert.attributeContains('button', 'aria-label', 'Custom Text')
+      .assert.containsText('button', '');
   },
 
   'Displays an initially open toggle-button': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/initially-open`)
-      .assert.attributeContains('.terra-Toggle', 'aria-hidden', 'false');
+      .assert.attributeContains('button + div', 'aria-hidden', 'false');
   },
 
   'Displays a toggle-button with customized openedButtonText': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/open-button-text`)
-      .click('.terra-Button')
-      .assert.containsText('.terra-ToggleButton-buttonText', 'Custom Text');
+      .click('button')
+      .assert.containsText('#openedButtonText button > div > div:nth-child(2) > span', 'Custom Text 2');
   },
 
   'Triggers onOpen callback for ToggleButton when button is selected': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/on-open-event`)
-      .click('.terra-Button')
+      .click('button')
       .assert.containsText('#on-open-event', 'Times Opened: 1')
-      .click('.terra-Button') // Close ToggleButton
-      .sendKeys('.terra-Button', browser.Keys.SPACE)
+      .click('button') // Close ToggleButton
+      .sendKeys('button', browser.Keys.SPACE)
       .assert.containsText('#on-open-event', 'Times Opened: 2')
-      .sendKeys('.terra-Button', browser.Keys.SPACE)  // Close ToggleButton
-      .sendKeys('.terra-Button', browser.Keys.ENTER)
+      .sendKeys('button', browser.Keys.SPACE)  // Close ToggleButton
+      .sendKeys('button', browser.Keys.ENTER)
       .assert.containsText('#on-open-event', 'Times Opened: 3');
   },
 
   'Triggers onClose callback for ToggleButton when button is selected': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/on-close-event`)
-      .click('.terra-Button') // Open ToggleButton
-      .click('.terra-Button')
+      .click('button') // Open ToggleButton
+      .click('button')
       .assert.containsText('#on-close-event', 'Times Closed: 1')
-      .click('.terra-Button') // Open ToggleButton
-      .sendKeys('.terra-Button', browser.Keys.SPACE)
+      .click('button') // Open ToggleButton
+      .sendKeys('button', browser.Keys.SPACE)
       .assert.containsText('#on-close-event', 'Times Closed: 2')
-      .sendKeys('.terra-Button', browser.Keys.SPACE) // Open ToggleButton
-      .sendKeys('.terra-Button', browser.Keys.ENTER)
+      .sendKeys('button', browser.Keys.SPACE) // Open ToggleButton
+      .sendKeys('button', browser.Keys.ENTER)
       .assert.containsText('#on-close-event', 'Times Closed: 3');
   },
 
@@ -98,8 +98,8 @@ module.exports = {
   'Display toggle-button content when button is clicked': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-button-tests/default`)
-      .click('.terra-Button')
-      .waitForElementPresent('.terra-Toggle', 1000)
-      .expect.element('.terra-Toggle').to.have.attribute('aria-hidden').to.equal('false');
+      .click('button')
+      .waitForElementPresent('button + div', 1000)
+      .expect.element('button + div').to.have.attribute('aria-hidden').to.equal('false');
   },
 };