+++ packages/terra-toggle/tests/nightwatch/toggle-spec.js
@@ -1,5 +1,5 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
-
+// eslint-disable-next-line import/no-extraneous-dependencies
 const screenshot = require('terra-toolkit').screenshot;
 
 module.exports = {
@@ -14,50 +14,47 @@ module.exports = {
   'Displays a default toggle': (browser) => {
     browser
       .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-tests/default`)
-      .assert.elementPresent('.terra-Toggle');
+      .assert.elementPresent('#toggle');
   },
 
   'Expanding the toggle: aria-hidden attribute': (browser) => {
     browser
      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-tests/default`)
-     .waitForElementPresent('.terra-Toggle', 1000)
-     .expect.element('.terra-Toggle').to.have.attribute('aria-hidden').to.equal('true');
+     .waitForElementPresent('#toggle', 1000)
+     .expect.element('#toggle').to.have.attribute('aria-hidden').to.equal('true');
 
     browser
      .click('button')
-     .waitForElementPresent('.terra-Toggle', 1000)
-     .expect.element('.terra-Toggle').to.have.attribute('aria-hidden').to.equal('false');
+     .waitForElementPresent('#toggle', 1000)
+     .expect.element('#toggle').to.have.attribute('aria-hidden').to.equal('false');
 
     browser
      .click('button')
-     .waitForElementPresent('.terra-Toggle', 1000)
-     .expect.element('.terra-Toggle').to.have.attribute('aria-hidden').to.equal('true');
+     .waitForElementPresent('#toggle', 1000)
+     .expect.element('#toggle').to.have.attribute('aria-hidden').to.equal('true');
   },
 
   'Render as opened': (browser) => {
     browser
      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-tests/open`)
-     .waitForElementPresent('.terra-Toggle', 1000)
-     .expect.element('.terra-Toggle').to.have.attribute('aria-hidden').to.equal('false');
+     .waitForElementPresent('#toggle', 1000)
+     .expect.element('#toggle').to.have.attribute('aria-hidden').to.equal('false');
   },
 
   'Disables focusable elements within animated toggle when closed': (browser) => {
     browser
      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-tests/animated`)
-     .assert.cssClassPresent('.terra-Toggle-content', 'is-closed')
-     .assert.cssProperty('.terra-Toggle-content.is-closed', 'visibility', 'hidden')
+     .assert.cssProperty('#toggle[aria-hidden="true"]', 'visibility', 'hidden')
      .click('button') // Opens toggle
+     .waitForElementPresent('#toggle[aria-hidden="false"]', 1000)
      .click('button') // Closes toggle
-     .waitForElementPresent('.terra-Toggle-content.is-closed', 1000)
-     .assert.cssClassPresent('.terra-Toggle-content', 'is-closed')
-     .assert.cssProperty('.terra-Toggle-content.is-closed', 'visibility', 'hidden');
+     .waitForElementPresent('#toggle[aria-hidden="true"]', 1000);
   },
 
   'Enables focusable elements within animated toggle when opened': (browser) => {
     browser
      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/toggle-tests/animated`)
      .click('button')
-     .assert.cssClassNotPresent('.terra-Toggle-content', 'is-closed')
-     .assert.cssProperty('.terra-Toggle-content', 'visibility', 'visible');
+     .assert.cssProperty('#toggle[aria-hidden="false"]', 'visibility', 'visible');
   },
 };