+++ packages/terra-badge/CHANGELOG.md
@@ -3,9 +3,15 @@ Changelog
 
 Unreleased
 ----------
+
+1.1.0 - (July 13, 2017)
+------------------
 ### Changed
 * Converted component to use CSS modules
 * Converted SCSS variables to CSS custom properties for theming
+* Normalized CSS custom properties naming
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
 
 1.0.0 - (June 28, 2017)
 ------------------