+++ packages/terra-modal/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Custom Props are now added to the modal, not the portal
+
+1.1.0 - (July 13, 2017)
+------------------
+Minor version bump
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
 
 1.0.0 - (June 28, 2017)
 ------------------