+++ packages/terra-time-input/CHANGELOG.md
@@ -3,8 +3,16 @@ Changelog
 
 Unreleased
 ----------
+
+1.1.0 - (July 13, 2017)
+------------------
 ### Changed
 * Converted component to use CSS modules
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
+
+### Fixed
+* Updated package.json to use rimraf
 
 1.0.0 - (June 28, 2017)
 ------------------